Fixes for logger getting cleaned up too early, also handles null terminated buffers properly.
